feat: Add Terraform code for creating EC2 instances and IAM users
This commit adds Terraform code to create EC2 instances and IAM users for the e-commerce component of the project. The EC2 instances will host the e-commerce application, and the IAM users will provide secure access to the necessary resources. These changes are crucial for the successful deployment and operation of the project.